% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bake.R
\name{bake}
\alias{bake}
\title{Leaf parameter temperature responses}
\usage{
bake(leaf_par, bake_par, constants, unitless = FALSE,
  check = !unitless)
}
\arguments{
\item{leaf_par}{A list of leaf parameters inheriting class \code{leaf_par}. This can be generated using the \code{make_leafpar} function.}

\item{bake_par}{A list of temperature response parameters inheriting class \code{bake_par}. This can be generated using the \code{make_bakepar} function.}

\item{constants}{A list of physical constants inheriting class \code{constants}. This can be generated using the \code{make_constants} function.}

\item{unitless}{Logical. Should units be set? The function is faster when FALSE, but input must be in correct units or else results will be incorrect without any warning.}

\item{check}{Logical. Should all parameter sets be checked? TRUE is safer, but FALSE is faster.}
}
\description{
a wrapper for \code{\link[photosynthesis]{bake}}
}
