% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_leaf.R
\name{carbon_balance}
\alias{carbon_balance}
\title{Calculate carbon balance}
\usage{
carbon_balance(T_leaf, carbon_costs, pars)
}
\arguments{
\item{T_leaf}{Leaf temperature in degree Kelvin with \code{units} dropped}

\item{carbon_costs}{A named list of resources with their costs in terms of carbon (e.g. mol C / mol H2O). Currently only H2O is supported. See details below.}

\item{pars}{Concatenated parameters (\code{leaf_par}, \code{enviro_par}, \code{bake_par}, and \code{constants}) with \code{units} dropped}
}
\value{
Value of class \code{numeric} indicating the carbon balance.
}
\description{
Calculate carbon balance
}
\details{
This function is not intended to be called directly because most checks have been removed for speed in \code{\link{optimize_leaf}}.

Currently only carbon a cost of water lost to transpiration is supported. Functions calculate the instantaneous leaf-level carbon gain and water loss per area. In the future, I plan to extend functionality to other resources (e.g. nitrogen) and integrate over time courses.

The basic equation is:

Carbon Balance = Carbon gain - (Carbon cost of water) Water loss
}
\examples{
library(magrittr)
cs <- make_constants()
lp <- make_leafpar(cs)
bp <- make_bakepar()  
ep <- make_enviropar()

T_leaf <- tealeaves::tleaf(lp, ep, cs, unitless = TRUE)$T_leaf
blp <- lp \%>\% 
  c(T_leaf = T_leaf) \%>\%
  bake(bp, cs)
carbon_costs <- list(H2O = 0.003)
pars <- c(cs, lp[!(names(lp) \%in\% names(blp))], blp, ep) \%>\%
  purrr::map_if(function(x) is(x, "units"), drop_units)
T_leaf \%<>\% drop_units()

# everything must be unitless
# pars must already be baked
carbon_balance(T_leaf, carbon_costs, pars)

}
