% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_Tleaf.R
\name{.get_Dx}
\alias{.get_Dx}
\title{D_x: Calculate diffusion coefficient for a given temperature and pressure}
\usage{
.get_Dx(D_0, Temp, eT, P)
}
\arguments{
\item{D_0}{Diffusion coefficient at 273.15 K and 101.3246 kPa}

\item{Temp}{Temperature in Kelvin}

\item{eT}{Exponent for temperature dependence of diffusion}

\item{P}{Atmospheric pressure in kPa}
}
\description{
D_x: Calculate diffusion coefficient for a given temperature and pressure
}
