% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_Tleaf.R
\name{.get_gbw}
\alias{.get_gbw}
\title{g_bw: Boundary layer conductance to water vapour}
\usage{
.get_gbw(T_leaf, surface, pars)
}
\arguments{
\item{T_leaf}{Leaf temperature in Kelvin}

\item{surface}{Leaf surface (lower or upper)}

\item{pars}{Concatenated parameters (\code{leaf_par}, \code{enviro_par}, and \code{constants})}
}
\description{
g_bw: Boundary layer conductance to water vapour
}
