% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_Tleaf.R
\name{.get_gh}
\alias{.get_gh}
\title{g_h: boundary layer conductance to heat (m s^-1)}
\usage{
.get_gh(T_leaf, surface, pars)
}
\arguments{
\item{T_leaf}{Leaf temperature in Kelvin}

\item{surface}{Leaf surface (lower or upper)}

\item{pars}{Concatenated parameters (\code{leaf_par}, \code{enviro_par}, and \code{constants})}
}
\description{
g_h: boundary layer conductance to heat (m s^-1)
}
