% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_Tleaf.R
\name{.get_gr}
\alias{.get_gr}
\title{Gr: Grashof number}
\usage{
.get_gr(T_leaf, pars)
}
\arguments{
\item{T_leaf}{Leaf temperature in Kelvin}

\item{pars}{Concatenated parameters (\code{leaf_par}, \code{enviro_par}, and \code{constants})}
}
\description{
Gr: Grashof number
}
