% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_Tleaf.R
\name{.get_L}
\alias{.get_L}
\title{L: Latent heat flux density (W m^-2)}
\usage{
.get_L(T_leaf, pars)
}
\arguments{
\item{T_leaf}{Leaf temperature in Kelvin}

\item{pars}{Concatenated parameters (\code{leaf_par}, \code{enviro_par}, and \code{constants})}
}
\description{
L: Latent heat flux density (W m^-2)
}
