% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_Tleaf.R
\name{.get_Pa}
\alias{.get_Pa}
\title{P_a: density of dry air (g m^-3)}
\usage{
.get_Pa(T_leaf, pars)
}
\arguments{
\item{T_leaf}{Leaf temperature in Kelvin}

\item{pars}{Concatenated parameters (\code{leaf_par}, \code{enviro_par}, and \code{constants})}
}
\description{
P_a: density of dry air (g m^-3)
}
