% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_Tleaf.R
\name{.get_ps}
\alias{.get_ps}
\title{Saturation water vapour pressure in kPa}
\usage{
.get_ps(Temp, P)
}
\arguments{
\item{Temp}{Temperature in Kelvin}

\item{P}{Atmospheric pressure in kPa}
}
\description{
Saturation water vapour pressure in kPa
}
