% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_Tleaf.R
\name{.get_Rabs}
\alias{.get_Rabs}
\title{R_abs: total absorbed radiation (W m^-2)}
\usage{
.get_Rabs(pars)
}
\arguments{
\item{pars}{Concatenated parameters (\code{leaf_par}, \code{enviro_par}, and \code{constants})}
}
\description{
R_abs: total absorbed radiation (W m^-2)
}
