% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_Tleaf.R
\name{.get_Rr}
\alias{.get_Rr}
\title{R_r: longwave re-radiation (W m^-2)}
\usage{
.get_Rr(pars)
}
\arguments{
\item{pars}{Concatenated parameters (\code{leaf_par}, \code{enviro_par}, and \code{constants})}
}
\description{
R_r: longwave re-radiation (W m^-2)
}
