% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_Tleaf.R
\name{.get_sh}
\alias{.get_sh}
\title{Sh: Sherwood number}
\usage{
.get_sh(T_leaf, surface, pars)
}
\arguments{
\item{T_leaf}{Leaf temperature in Kelvin}

\item{surface}{Leaf surface (lower or upper)}

\item{pars}{Concatenated parameters (\code{leaf_par}, \code{enviro_par}, and \code{constants})}
}
\description{
Sh: Sherwood number
}
