% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_Tleaf.R
\name{.get_Tv}
\alias{.get_Tv}
\title{Calculate virtual temperature}
\usage{
.get_Tv(Temp, p_air, P)
}
\arguments{
\item{Temp}{Temperature in Kelvin}

\item{p_air}{Saturation water vapour pressure of air in kPa}

\item{P}{Atmospheric pressure in kPa}
}
\description{
Calculate virtual temperature
}
