% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_parameters.R
\name{.missing_traits}
\alias{.missing_traits}
\title{Traits missing from traits argument (i.e. traits to be fixed rather than optimized)}
\usage{
.missing_traits(leaf_traits)
}
\arguments{
\item{leaf_traits}{A vector of leaf traits to optimize. Must one or more of stomatal conductance (\code{g_sw}), stomatal ratio (\code{sr}), and leaf size (\code{leafsize}). Partial matching allowed.}
}
\description{
Traits missing from traits argument (i.e. traits to be fixed rather than optimized)
}
