% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve_leaf.R
\name{evolve_leaf}
\alias{evolve_leaf}
\title{Find evolutionarily optimal leaf traits}
\usage{
evolve_leaf(leaf_traits, constraints, leaf_par, enviro_par, constants, ...)
}
\arguments{
\item{leaf_traits}{A vector of leaf traits to optimize. Must one or more of stomatal conductance (\code{g_sw}), stomatal ratio (\code{sr}), and leaf size (\code{leafsize}). Partial matching allowed.}

\item{constraints}{A list containing lambda (i.e. the cost of water), which must be a scalar. No other constraints are currently implemented.}

\item{leaf_par}{A list of leaf parameters. This can be generated using the \code{make_leafpar} function.}

\item{enviro_par}{A list of environmental parameters. This can be generated using the \code{make_enviropar} function.}

\item{constants}{A list of physical constants. This can be generated using the \code{make_constants} function.}

\item{...}{Additional arguments passed to \code{optim}}
}
\description{
Find evolutionarily optimal leaf traits
}
