% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_Tleaf.R
\name{find_Tleaf}
\alias{find_Tleaf}
\title{Find leaf temperature}
\usage{
find_Tleaf(leaf_par, enviro_par, constants)
}
\arguments{
\item{leaf_par}{A list of leaf parameters. This can be generated using the \code{make_leafpar} function.}

\item{enviro_par}{A list of environmental parameters. This can be generated using the \code{make_enviropar} function.}

\item{constants}{A list of physical constants. This can be generated using the \code{make_constants} function.}
}
\description{
Find leaf temperature
}
