% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_fitness.R
\name{leaf_fitness}
\alias{leaf_fitness}
\title{Leaf fitness}
\usage{
leaf_fitness(leaf_traits, constraints, leaf_par, enviro_par, constants,
  return_negative = FALSE, ...)
}
\arguments{
\item{leaf_traits}{A vector of leaf traits to optimize. Must one or more of stomatal conductance (\code{g_sw}), stomatal ratio (\code{sr}), and leaf size (\code{leafsize}). Partial matching allowed.}

\item{constraints}{A list containing lambda (i.e. the cost of water), which must be a scalar. No other constraints are currently implemented.}

\item{leaf_par}{A list of leaf parameters. This can be generated using the \code{make_leafpar} function.}

\item{enviro_par}{A list of environmental parameters. This can be generated using the \code{make_enviropar} function.}

\item{constants}{A list of physical constants. This can be generated using the \code{make_constants} function.}

\item{return_negative}{Logical. Should negative fitness be returned? This is useful for optimization, which usually search for minimum.}

\item{...}{ignored}
}
\description{
Leaf fitness
}
