% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafoptimizer.R
\docType{package}
\name{leafoptimizer}
\alias{leafoptimizer}
\alias{leafoptimizer-package}
\title{\code{leafoptimizer} package}
\description{
Optimize leaf traits to different environments in silico
}
\details{
See the README on
\href{https://github.com/cdmuir/leafoptimizer}{GitHub}
}
