% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_leaf.R
\encoding{UTF-8}
\name{optimize_leaves}
\alias{optimize_leaves}
\alias{optimize_leaf}
\title{Optimize leaf photosynthesis}
\usage{
optimize_leaves(traits, carbon_costs, leaf_par, enviro_par, bake_par,
  constants, progress = TRUE, quiet = FALSE)

optimize_leaf(traits, carbon_costs, leaf_par, enviro_par, bake_par,
  constants, quiet = FALSE)
}
\arguments{
\item{traits}{A vector of one or more character strings indicating which trait(s) to optimize. Stomatal conductance (\code{g_sc}) and stomatal ratio (\code{logit_sr}) are currently supported.}

\item{carbon_costs}{A named list of resources with their costs in terms of carbon (e.g. mol C / mol H2O). Currently only H2O is supported. See details below.}

\item{leaf_par}{A list of leaf parameters inheriting class \code{leaf_par}. This can be generated using the \code{make_leafpar} function.}

\item{enviro_par}{A list of environmental parameters inheriting class \code{enviro_par}. This can be generated using the \code{make_enviropar} function.}

\item{bake_par}{A list of temperature response parameters inheriting class \code{bake_par}. This can be generated using the \code{make_bakepar} function.}

\item{constants}{A list of physical constants inheriting class \code{constants}. This can be generated using the \code{make_constants} function.}

\item{progress}{Logical. Should a progress bar be displayed?}

\item{quiet}{Logical. Should messages be displayed?}
}
\value{
A data.frame with the following \code{units} columns \cr

\tabular{ll}{

\bold{Input:} \tab \cr
\cr
\code{C_air} \tab atmospheric CO2 concentration (Pa) \cr
\code{g_mc25} \tab mesophyll conductance to CO2 at 25 °C (\eqn{\mu}mol CO2 / (m\eqn{^2} s Pa)) \cr
\code{g_sc} \tab stomatal conductance to CO2 (\eqn{\mu}mol CO2 / (m\eqn{^2} s Pa)) \cr
\code{g_uc} \tab cuticular conductance to CO2 (\eqn{\mu}mol CO2 / (m\eqn{^2} s Pa)) \cr
\code{gamma_star25} \tab chloroplastic CO2 compensation point at 25 °C (Pa) \cr
\code{J_max25} \tab potential electron transport at 25 °C (\eqn{\mu}mol CO2) / (m\eqn{^2} s) \cr
\code{K_C25} \tab Michaelis constant for carboxylation at 25 °C (\eqn{\mu}mol / mol) \cr
\code{K_O25} \tab Michaelis constant for oxygenation at 25 °C (\eqn{\mu}mol / mol) \cr
\code{k_mc} \tab partition of \eqn{g_\mathrm{mc}}{g_mc} to lower mesophyll (unitless) \cr
\code{k_sc} \tab partition of \eqn{g_\mathrm{sc}}{g_sc} to lower surface (unitless) \cr
\code{k_uc} \tab partition of \eqn{g_\mathrm{uc}}{g_uc} to lower surface (unitless) \cr
\code{leafsize} \tab leaf characteristic dimension (m) \cr
\code{O} \tab atmospheric O2 concentration (kPa) \cr
\code{P} \tab atmospheric pressure (kPa) \cr
\code{phi} \tab initial slope of the response of J to PPFD (unitless) \cr
\code{PPFD} \tab photosynthetic photon flux density (umol quanta / (m\eqn{^2} s)) \cr
\code{R_d25} \tab nonphotorespiratory CO2 release  at 25 °C (\eqn{\mu}mol CO2 / (m\eqn{^2} s)) \cr
\code{RH} \tab relative humidity (unitless) \cr
\code{theta_J} \tab curvature factor for light-response curve (unitless) \cr
\code{T_air} \tab air temperature (K) \cr
\code{T_leaf} \tab leaf tempearture (K) \cr
\code{V_cmax25} \tab maximum rate of carboxylation at 25 °C (\eqn{\mu}mol CO2 / (m\eqn{^2} s)) \cr
\code{V_tpu25} \tab rate of triose phosphate utilisation at 25 °C (\eqn{\mu}mol CO2 / (m\eqn{^2} s)) \cr
\code{wind} \tab wind speed (m / s) \cr
\cr
\bold{Baked Input:} \tab \cr
\cr
\code{g_mc} \tab mesophyll conductance to CO2 at \code{T_leaf} (\eqn{\mu}mol CO2 / (m\eqn{^2} s Pa)) \cr
\code{gamma_star} \tab chloroplastic CO2 compensation point at \code{T_leaf} (Pa) \cr
\code{J_max} \tab potential electron transport at \code{T_leaf} (\eqn{\mu}mol CO2) / (m\eqn{^2} s) \cr
\code{K_C} \tab Michaelis constant for carboxylation at \code{T_leaf} (\eqn{\mu}mol / mol) \cr
\code{K_O} \tab Michaelis constant for oxygenation at \code{T_leaf}(\eqn{\mu}mol / mol) \cr
\code{R_d} \tab nonphotorespiratory CO2 release  at \code{T_leaf} (\eqn{\mu}mol CO2 / (m\eqn{^2} s)) \cr
\code{V_cmax} \tab maximum rate of carboxylation at \code{T_leaf} (\eqn{\mu}mol CO2 / (m\eqn{^2} s)) \cr
\code{V_tpu} \tab rate of triose phosphate utilisation at \code{T_leaf} (\eqn{\mu}mol CO2 / (m\eqn{^2} s)) \cr
\cr
\bold{Output:} \tab \cr
\cr
\code{A} \tab photosynthetic rate at \code{C_chl} (\eqn{\mu}mol CO2 / (m\eqn{^2} s)) \cr
\code{C_chl} \tab chloroplastic CO2 concentration where \code{A_supply} intersects \code{A_demand} (Pa) \cr
\code{g_tc} \tab total conductance to CO2 at \code{T_leaf} (\eqn{\mu}mol CO2 / (m\eqn{^2} s Pa)) \cr
\code{value} \tab \code{A_supply} - \code{A_demand} (\eqn{\mu}mol CO2 / (m\eqn{^2} s)) at \code{C_chl} \cr
\code{convergence} \tab convergence code (0 = converged)
}
}
\description{
\code{optimize_leaves}: optimize C3 photosynthesis over multiple parameter sets

\code{optimize_leaf}: simulate C3 photosynthesis over a single parameter set
}
\details{
\code{optimize_leaf}: This function optimizes leaf traits using an integrated leaf temperature and C3 photosynthesis model under a set of environmental conditions. The leaf temperature model is described in the \code{\link[tealeaves]{tealeaves}} package. The C3 photosynthesis model is described in the \code{\link[photosynthesis]{photosynthesis-package}} package\cr
\cr
\code{optimize_leaves}: This function uses \code{optimize_leaf} to optimize over multiple parameter sets that are generated using \code{\link[tidyr]{crossing}}. \cr
}
\examples{
# Single parameter set with 'photo'

constants <- make_constants()
leaf_par <- make_leafpar(constants)
enviro_par <- make_enviropar()
bake_par <- make_bakepar()
# optimize_leaf(leaf_par, enviro_par, bake_par, constants)

# Multiple parameter sets with 'photosynthesis'

leaf_par <- make_leafpar(constants,
  replace = list(
    T_leaf = set_units(c(293.14, 298.15), "K")
    )
  )
# optimize_leaves(leaf_par, enviro_par, bake_par, constants)

}
