% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sun2ppfd}
\alias{sun2ppfd}
\alias{ppfd2sun}
\title{Convert irradiance (W/m^2) to PPFD (umol/m^2/s)}
\usage{
sun2ppfd(S_sw, f_par, E_q)

ppfd2sun(PPFD, f_par, E_q)
}
\arguments{
\item{S_sw}{incident short-wave (solar) radiation flux density with units (W / m^2) of class \code{units}.}

\item{f_par}{fraction of \code{S_sw} that is photosynthetic active radiation (PAR). Must be unitless value between 0 and 1.}

\item{E_q}{energy per mole quanta. Value in kJ/mol of class \code{units}.}

\item{PPFD}{Photosynthetic photon flux density in \eqn{\mu}mol quanta / (m^2 s) of class \code{units}.}
}
\value{
Value with units \eqn{\mu}mol / (m^2 s) of class \code{units}.
}
\description{
Convert irradiance (W/m^2) to PPFD (umol/m^2/s)

Convert PPFD (umol/m^2/s) to irradiance (W/m^2)
}
\details{
Shortwave radiation is (at first approximation) the sum of photosynthetically active radiation (PAR) and near-infrared radiation (NIR):

\deqn{S_\mathrm{sw} = S_\mathrm{PAR} + S_\mathrm{NIR}}{S_sw = S_par + S_nir}

Most sources (e.g. Jones 2013) assume that S_nir = S_par for sunlight, so \eqn{f_\mathrm{PAR} = 0.5}{f_par = 0.5}.

To convert PAR to PPFD, divide by the energy per mol quanta. Gutschick (2016) suggests ~220 kJ/mol quanta for PAR:

\deqn{\mathrm{PPFD} = S_\mathrm{PAR} / E_q}{PPFD = S_par / E_q}
}
\examples{
S_sw <- set_units(1000, "W/m^2")
f_par <- set_units(0.5)
E_q <- set_units(220, "kJ/mol")
sun2ppfd(S_sw, f_par, E_q)

}
\references{
Jones HG. 2013. Plants and microclimate: a quantitative approach to environmental plant physiology. Cambridge University Press.

Gutschick VP. 2016. Leaf energy balance: basics, and modeling from leaves to Canopies. In Canopy Photosynthesis: From Basics to Applications (pp. 23-58). Springer, Dordrecht.
}
