% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_leaf.R
\name{carbon_balance}
\alias{carbon_balance}
\title{Calculate carbon balance}
\usage{
carbon_balance(T_leaf, carbon_costs, pars)
}
\arguments{
\item{T_leaf}{Leaf temperature in degree Kelvin}

\item{carbon_costs}{A named list of resources with their costs in terms of carbon (e.g. mol C / mol H2O). Currently only H2O is supported. See details below.}

\item{pars}{Concatenated parameters (\code{leaf_par}, \code{enviro_par}, \code{bake_par}, and \code{constants})
#' @param check Logical. Should all parameter sets be checked? TRUE is safer, but FALSE is faster.}
}
\value{
Value of class \code{units} indicating the carbon balance.
}
\description{
Calculate carbon balance
}
\details{
Currently only carbon a cost of water lost to transpiration is supported. Functions calculate the instantaneous leaf-level carbon gain and water loss per area. In the future, I plan to extend functionality to other resources (e.g. nitrogen) and integrate over time courses.

The basic equation is:

Carbon Balance = Carbon gain - (Carbon cost of water) Water loss
}
\examples{
cs <- make_constants()
lp <- make_leafpar(cs)
bp <- make_bakepar()  
ep <- make_enviropar()

T_leaf <- tealeaves::tleaf(lp, ep, cs)$T_leaf
carbon_costs <- list(H2O = 0.003)
pars <- c(cs, lp, bp, ep)

carbon_balance(T_leaf, carbon_costs, pars)

}
