% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafevolver.R
\docType{package}
\name{leafevolver}
\alias{leafevolver}
\alias{leafevolver-package}
\title{\code{leafevolver} package}
\description{
Evolve leaf traits to different environments in silico
}
\details{
See the README on
\href{https://github.com/cdmuir/leafevolver}{GitHub}
}
